+++ fearless-utils/src/test/java/jp/co/soramitsu/fearless_utils/encrypt/LocalSubstrateKeypairDerivationTest.kt
@@ -1,13 +1,13 @@
-package jp.co.soramitsu.fearless_utils.encrypt.seed.substrate
+package jp.co.soramitsu.fearless_utils.encrypt
 
 import jp.co.soramitsu.fearless_utils.encrypt.EncryptionType
-import jp.co.soramitsu.fearless_utils.encrypt.mnemonic.SubstrateSeedFactoryTest
+import jp.co.soramitsu.fearless_utils.encrypt.SubstrateKeypairDerivationTest
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.junit.MockitoJUnitRunner
 
 @RunWith(MockitoJUnitRunner::class)
-class LocalSubstrateSeedFactoryTest : SubstrateSeedFactoryTest() {
+class LocalSubstrateKeypairDerivationTest : SubstrateKeypairDerivationTest() {
 
     @Test
     fun `should  pass ed25519 tests`() {
