+++ fearless-utils/src/testShared/java/jp/co/soramitsu/fearless_utils/encrypt/SubstrateKeypairDerivationTest.kt
@@ -1,15 +1,15 @@
-package jp.co.soramitsu.fearless_utils.encrypt.mnemonic
+package jp.co.soramitsu.fearless_utils.encrypt
 
 import com.google.gson.Gson
-import jp.co.soramitsu.fearless_utils.encrypt.EncryptionType
 import jp.co.soramitsu.fearless_utils.encrypt.keypair.substrate.SubstrateKeypairFactory
 import jp.co.soramitsu.fearless_utils.extensions.toHexString
 import jp.co.soramitsu.fearless_utils.getResourceReader
 import jp.co.soramitsu.fearless_utils.encrypt.junction.SubstrateJunctionDecoder
+import jp.co.soramitsu.fearless_utils.encrypt.mnemonic.MnemonicTestCase
 import jp.co.soramitsu.fearless_utils.encrypt.seed.substrate.SubstrateSeedFactory
 import org.junit.Assert
 
-abstract class SubstrateSeedFactoryTest {
+abstract class SubstrateKeypairDerivationTest {
 
     val gson = Gson()
